requirements.txt
absl-py==1.3.0
apturl==0.5.2
asttokens==2.0.8
astunparse==1.6.3
attrs==21.4.0
autopep8==2.0.0
backcall==0.2.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==5.2.0
catkin-pkg==0.4.24
catkin-pkg-modules==0.5.2
catkin-tools==0.8.2
catkin-tools-fetch==0.3.5
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
commonmark==0.9.1
contourpy==1.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.18.1
duplicity==0.8.12.0
einops==0.6.0
empy==3.3.4
entrypoints==0.3
et-xmlfile==1.1.0
executing==1.1.0
fasteners==0.14.1
flatbuffers==22.9.24
fonttools==4.28.3
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.4.0
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
httplib2==0.14.0
idna==2.8
importlib-metadata==4.10.1
iopath==0.1.10
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
jsonschema==3.2.0
jupyter-client==7.3.5
jupyter-core==4.11.1
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==14.0.6
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.0
matplotlib-inline==0.1.6
meld==3.20.2
mock==4.0.3
monotonic==1.5
more-itertools==4.2.0
nest-asyncio==1.5.5
netifaces==0.10.4
numpy==1.21.4
oauthlib==3.1.0
olefile==0.46
openai==0.23.1
opencv-python==4.5.4.60
openpyxl==3.0.10
opt-einsum==3.3.0
osrf-pycommon==2.0.0
packaging==21.3
pandas==1.5.0
pandas-stubs==1.5.0.221012
paramiko==2.6.0
parso==0.8.3
pdf2image==1.16.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.4.0
portalocker==2.6.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.9.1
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.13.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.6
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
pytz==2022.4
pyxdg==0.26
PyYAML==5.3.1
pyzmq==24.0.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rich==12.6.0
roman==2.0.0
rsa==4.9
scipy==1.9.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
# Editable install with no version control (semseg==0.4.1)
-e /home/shankhanil/work/3dclothgen/pipeline/semantic-segmentation
setuptools-scm==6.3.2
simplejson==3.16.0
sip==4.19.21
six==1.16.0
stack-data==0.5.1
systemd-python==234
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==1.1.0
texttable==1.6.4
tomli==1.2.2
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
types-pytz==2022.4.0.0
typing-extensions==4.3.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
websocket-client==0.59.0
Werkzeug==2.2.2
wrapt==1.14.1
xkit==0.0.0
yacs==0.1.8
zipp==3.7.0
